deprecate JRubyPrepareGems in favour of JRubyPrepare
