+++ jruby-gradle-base-plugin/src/main/groovy/com/github/jrubygradle/JRubyPlugin.groovy
@@ -4,6 +4,7 @@ import com.github.jrubygradle.internal.JRubyExecDelegate
 import com.github.jrubygradle.internal.GemVersionResolver
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.Task
 import org.gradle.api.tasks.Copy
 import org.gradle.api.tasks.Delete
 import org.gradle.api.tasks.bundling.War
@@ -26,7 +27,6 @@ class JRubyPlugin implements Plugin<Project> {
 
         // Set up a special configuration group for our embedding jars
         project.configurations.create('gems')
-        project.configurations.create('jrubyEmbeds')
         project.configurations.create(JRubyExec.JRUBYEXEC_CONFIG)
         JRubyExecDelegate.addToProject(project)
 
@@ -41,22 +41,20 @@ class JRubyPlugin implements Plugin<Project> {
             JRubyExec.updateJRubyDependencies(project)
         }
 
-        project.task('jrubyPrepareGems', type: JRubyPrepareGems) {
-            logger.info 'Deprecated use "jrubyPrepare" instead.'
-            group TASK_GROUP_NAME
-            description 'Prepare the gems/jars from the `gem` dependencies, extracts the gems into jruby.installGemDir and sets up the jars in jruby.installGemDir/jars'
-            gems project.configurations.gems
-            outputDir project.jruby.gemInstallDir
-        }
+        Task jrpg = project.tasks.create 'jrubyPrepareGems'
+        jrpg.dependsOn 'jrubyPrepare'
+        jrpg << { logger.info "'jrubyPrepareGems' is deprecated and will be removed in a future version. Use 'jrubyPrepare' instead." }
+        jrpg.group 'Deprecated'
+        jrpg.description "Prepare the gems/jars from the `gem` dependencies, extracts the gems into jruby.installGemDir and sets up the jars in jruby.installGemDir/jars"
 
         project.task('jrubyPrepareJars', type: JRubyPrepareJars) {
             logger.info 'Obsolete tasks - does nothing anymore.'
             outputDir project.jruby.gemInstallDir
         }
 
-        project.task('jrubyPrepare', type: JRubyPrepareGems) {
+        project.task('jrubyPrepare', type: JRubyPrepare) {
             group TASK_GROUP_NAME
-            description 'Prepare the gems from the `gem` dependencies, extracts into jruby.installGemDir'
+            description 'Prepare the gems/jars from the `gem` dependencies, extracts the gems into jruby.installGemDir and sets up the jars in jruby.installGemDir/jars'
             gems project.configurations.gems
             outputDir project.jruby.gemInstallDir
         }
@@ -64,7 +62,7 @@ class JRubyPlugin implements Plugin<Project> {
         project.task('jrubyGenerateGradleRb', type: GenerateGradleRb) {
             group TASK_GROUP_NAME
             description 'Generate a gradle.rb stub for executing Ruby binstubs'
-            dependsOn project.tasks.jrubyPrepareGems
+            dependsOn project.tasks.jrubyPrepare
         }
     }
 }
