+++ jruby-gradle-base-plugin/src/main/groovy/com/github/jrubygradle/JRubyPrepareGems.groovy
@@ -13,52 +13,13 @@ import org.gradle.api.tasks.TaskAction
  * @author Schalk W. Cronjé
  * @author R Tyler Croy
  */
-class JRubyPrepareGems  extends DefaultTask {
+@Deprecated
+class JRubyPrepareGems  extends JRubyPrepare {
 
-    /** Target directory for GEMs. Extracted GEMs should end up in {@code outputDir + "/gems"}
-     */
-    @OutputDirectory
-    File outputDir
-
-
-    @InputFiles
-    FileCollection getGems() {
-        GemUtils.getGems(project.files(this.gems))
+  JRubyPrepareGems () {
+        super()
+        project.logger.info "The 'JRubyPrepareGems' task type is deprecated and will be removed in a future version. Please use 'JRubyPrepare' tasl type instead."
     }
 
-    /** Sets the output directory
-     *
-     * @param f Output directory
-     */
-    void outputDir(Object f) {
-        this.outputDir = project.file(f)
-    }
-
-    /** Adds gems to be prepared
-     *
-     * @param f A file, directory, configuration or list of gems
-     */
-    void gems(Object f) {
-        if (this.gems == null) {
-            this.gems = []
-        }
-        this.gems.add(f)
-    }
-
-    @TaskAction
-    void copy() {
-        File jrubyJar = JRubyExecUtils.jrubyJar(project.configurations.getByName(JRubyExec.JRUBYEXEC_CONFIG))
-        GemUtils.extractGems(project, jrubyJar, getGems(), outputDir, GemUtils.OverwriteAction.SKIP)
-        
-        if (gems != null) {
-            gems.each {
-                if (it instanceof Configuration) {
-                    GemUtils.setupJars(it, outputDir, GemUtils.OverwriteAction.SKIP)
-                }
-            }
-        }
-    }
-
-    private List<Object> gems
 }
